chore(wizard-refactor): extract resolvers/institution-by-name
ResolvedCourse + ResolvedSubject + ResolvedInstitution + resolveInstitutionByName.
Last of the four resolvers. Dispatcher 11/11 + lint 17/17 green.
